Duzo zmian, glownie w residual
Przede wszystkim dodano funkcje cheapest, ktora generuje najtansza droge ktora daje maxFlow, modDane uwzglednia droge, sporo funkcji w residual ktore byly konieczne, moga sie jeszcze przydac i moze dzialaja poprawnie. operator=, con kopiujacy, prosty druk, destruktor, usuwanie krawedzi

+++ BardzoWIP/zmodyfikujdane.h
@@ -14,8 +14,8 @@ ile pol, (nr. pola, prod)
 ile browarow, (nr, poj/prod)
 ile tawern (nr, poj)
 ile zwyklych skrzyzowan*/
-class zmodyfikujDane{
+class zmodyfikujDane {
 public:
-std::vector<int> modDane(std::string& filePath, int mode);
+	std::vector<float> modDane(std::string& filePath, int mode);
 };
 #endif
